+++ s2fft/general_precompute/resampling_jax.py
@@ -23,7 +23,7 @@ def mw_to_mwss_theta(f_mw: jnp.ndarray, L: int, spin: int = 0) -> jnp.ndarray:
 
     f_mw_ext = periodic_extension(f_mw, L, spin=spin, sampling="mw")
     fmp_mwss_ext = jnp.zeros(
-        (f_mw_ext.shape[0], 2 * L, 2 * L - 1), dtype=jnp.complex64
+        (f_mw_ext.shape[0], 2 * L, 2 * L - 1), dtype=jnp.complex128
     )
 
     fmp_mwss_ext = fmp_mwss_ext.at[:, 1:, :].set(
@@ -54,7 +54,7 @@ def mw_to_mwss_theta(f_mw: jnp.ndarray, L: int, spin: int = 0) -> jnp.ndarray:
 
 @partial(jit, static_argnums=(1))
 def mw_to_mwss_phi(f_mw: jnp.ndarray, L: int) -> jnp.ndarray:
-    f_mwss = jnp.zeros((f_mw.shape[0], L + 1, 2 * L), dtype=jnp.complex64)
+    f_mwss = jnp.zeros((f_mw.shape[0], L + 1, 2 * L), dtype=jnp.complex128)
     f_mwss = f_mwss.at[:, :, 1:].set(
         jnp.fft.fftshift(jnp.fft.fft(f_mw, axis=-1, norm="forward"), axes=-1)
     )
@@ -77,7 +77,7 @@ def periodic_extension(
     ntheta_ext = samples.ntheta_extension(L, sampling)
     m_offset = 1 if sampling == "mwss" else 0
 
-    f_ext = jnp.zeros((f.shape[0], ntheta_ext, nphi), dtype=jnp.complex64)
+    f_ext = jnp.zeros((f.shape[0], ntheta_ext, nphi), dtype=jnp.complex128)
     f_ext = f_ext.at[:, 0:ntheta, 0:nphi].set(f[:, 0:ntheta, 0:nphi])
     f_ext = jnp.fft.fftshift(
         jnp.fft.fft(f_ext, axis=-1, norm="backward"), axes=-1
@@ -175,7 +175,7 @@ def upsample_by_two_mwss_ext(f_ext: jnp.ndarray, L: int) -> jnp.ndarray:
 
     ntheta_ext_up = 2 * ntheta_ext
     f_ext_up = jnp.zeros(
-        (f_ext.shape[0], ntheta_ext_up, nphi), dtype=jnp.complex64
+        (f_ext.shape[0], ntheta_ext_up, nphi), dtype=jnp.complex128
     )
     f_ext_up = f_ext_up.at[:, L : ntheta_ext + L, :nphi].set(
         f_ext[:, 0:ntheta_ext, :nphi]
@@ -198,7 +198,7 @@ def periodic_extension_spatial_mwss(
     nphi = 2 * L
     ntheta_ext = 2 * L
 
-    f_ext = jnp.zeros((f.shape[0], ntheta_ext, nphi), dtype=jnp.complex64)
+    f_ext = jnp.zeros((f.shape[0], ntheta_ext, nphi), dtype=jnp.complex128)
     f_ext = f_ext.at[:, 0:ntheta, 0:nphi].set(f[:, 0:ntheta, 0:nphi])
     if spin.size > 1:
         f_ext = f_ext.at[:, ntheta:, 0 : 2 * L].set(
