+++ s2fft/general_precompute/spin_spherical.py
@@ -104,7 +104,7 @@ def inverse_transform(
     m_offset = 1 if sampling in ["mwss", "healpix"] else 0
     m_start_ind = L - 1 if reality else 0
 
-    ftm = np.zeros(samples.ftm_shape(L, sampling, nside), dtype=np.complex64)
+    ftm = np.zeros(samples.ftm_shape(L, sampling, nside), dtype=np.complex128)
     ftm[:, m_start_ind + m_offset :] = np.einsum(
         "...tlm, ...lm -> ...tm", kernel, flm[:, m_start_ind:]
     )
@@ -168,7 +168,7 @@ def inverse_transform_jax(
     m_offset = 1 if sampling in ["mwss", "healpix"] else 0
     m_start_ind = L - 1 if reality else 0
 
-    ftm = jnp.zeros(samples.ftm_shape(L, sampling, nside), dtype=jnp.complex64)
+    ftm = jnp.zeros(samples.ftm_shape(L, sampling, nside), dtype=jnp.complex128)
     ftm = ftm.at[:, m_start_ind + m_offset :].add(
         jnp.einsum(
             "...tlm, ...lm -> ...tm",
@@ -314,7 +314,7 @@ def forward_transform(
         else:
             ftm = np.fft.fft(f, axis=-1, norm="backward")
             ftm = np.fft.fftshift(ftm, axes=-1)[:, m_offset:]
-    flm = np.zeros(samples.flm_shape(L), dtype=np.complex64)
+    flm = np.zeros(samples.flm_shape(L), dtype=np.complex128)
     flm[:, m_start_ind:] = np.einsum("...tlm, ...tm -> ...lm", kernel, ftm)
 
     if reality:
@@ -383,7 +383,7 @@ def forward_transform_jax(
             ftm = jnp.fft.fft(f, axis=-1, norm="backward")
             ftm = jnp.fft.fftshift(ftm, axes=-1)[:, m_offset:]
 
-    flm = jnp.zeros(samples.flm_shape(L), dtype=jnp.complex64)
+    flm = jnp.zeros(samples.flm_shape(L), dtype=jnp.complex128)
     flm = flm.at[:, m_start_ind:].set(
         jnp.einsum("...tlm, ...tm -> ...lm", kernel, ftm, optimize=True)
     )
